requirements.txt
Django==2.0.3
django-filter==1.1.0
djangorestframework==3.7.7
python-nmap==0.6.1
